requirements.txt
fastapi 
uvicorn 
jinja2 
aiohttp 
azure-cosmos
python-dotenv